+++ sshmitm/apps/mockserver/__init__.py
@@ -1,0 +1,3 @@
+from sshmitm.apps.mockserver.mockserver import MockServer
+
+__all__ = ["MockServer"]